requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
asyncpg==0.29.0
asyncssh==2.16.0
bcrypt==4.2.0
certifi==2024.7.4
cffi==1.17.0
click==8.1.7
colorama==0.4.6
cryptography==43.0.0
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.112.2
fastapi-cache2==0.2.2
fastapi-cli==0.0.5
greenlet==3.0.3
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
orjson==3.10.7
packaging==24.1
paramiko==3.4.1
pendulum==3.0.0
psycopg-binary==3.2.1
psycopg2==2.9.9
pycparser==2.22
pydantic==2.8.2
pydantic-extra-types==2.9.0
pydantic-settings==2.4.0
pydantic_core==2.20.1
Pygments==2.18.0
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pywin32==306; platform_system=='Windows'
PyYAML==6.0.2
redis==4.6.0
rich==13.8.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
sshtunnel==0.4.0
starlette==0.38.2
time-machine==2.15.0
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
uvicorn==0.30.6
watchfiles==0.24.0
websockets==13.0.1
sshtunnel
cachetools